+++ src/rxn_network/core/cost_function.py
@@ -1,0 +1,12 @@
+from abc import ABCMeta, abstractmethod
+from monty.json import MSONable
+from rxn_network.core.reaction import Reaction
+
+
+class CostFunction(MSONable, metaclass=ABCMeta):
+    " Base definition for a cost function "
+
+    @abstractmethod
+    def evaluate(self, rxn: Reaction) -> float:
+        " Evaluates the total cost function on a reaction "
+
