+++ src/rxn_network/enumerators/utils.py
@@ -7,7 +7,7 @@ from rxn_network.reactions.computed import ComputedReaction
 from rxn_network.reactions.open import OpenComputedReaction
 
 
-def initialize_entry(formula, entry_set):
+def initialize_entry(formula, entry_set, stabilize=True):
     """
 
     Args:
@@ -17,7 +17,9 @@ def initialize_entry(formula, entry_set):
     Returns:
 
     """
-    entry = entry_set.stabilize_entry(entry_set.get_min_entry_by_formula(formula))
+    entry = entry_set.get_min_entry_by_formula(formula)
+    if stabilize:
+        entry = entry_set.stabilize_entry(entry)
     return entry
