+++ src/rxn_network/pathways/balanced.py
@@ -4,7 +4,8 @@ from typing import List, Optional, Union
 
 import numpy as np
 
-from rxn_network.core import Pathway, Reaction
+from rxn_network.core import Pathway
+from rxn_network.core import Reaction
 from rxn_network.reactions.basic import BasicReaction
 from rxn_network.pathways.basic import BasicPathway
 from rxn_network.pathways.utils import balance_path_arrays
